removing bootstrap
