--- cmd/bootstrap/main.go
@@ -1,160 +1,0 @@
-// Copyright (C) 2019 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-package main
-
-import (
-	"fmt"
-	"os"
-	"path/filepath"
-
-	"github.com/spf13/cobra"
-	"github.com/zeebo/errs"
-	"go.uber.org/zap"
-
-	"storj.io/storj/bootstrap"
-	"storj.io/storj/bootstrap/bootstrapdb"
-	"storj.io/storj/internal/fpath"
-	"storj.io/storj/internal/version"
-	"storj.io/storj/pkg/cfgstruct"
-	"storj.io/storj/pkg/process"
-	"storj.io/storj/pkg/revocation"
-)
-
-var (
-	rootCmd = &cobra.Command{
-		Use:   "bootstrap",
-		Short: "bootstrap",
-	}
-	runCmd = &cobra.Command{
-		Use:   "run",
-		Short: "Run the bootstrap server",
-		RunE:  cmdRun,
-	}
-	setupCmd = &cobra.Command{
-		Use:         "setup",
-		Short:       "Create config files",
-		RunE:        cmdSetup,
-		Annotations: map[string]string{"type": "setup"},
-	}
-
-	runCfg   bootstrap.Config
-	setupCfg bootstrap.Config
-
-	confDir     string
-	identityDir string
-)
-
-const (
-	defaultServerAddr = ":28967"
-)
-
-func init() {
-	defaultConfDir := fpath.ApplicationDir("storj", "bootstrap")
-	defaultIdentityDir := fpath.ApplicationDir("storj", "identity", "bootstrap")
-	cfgstruct.SetupFlag(zap.L(), rootCmd, &confDir, "config-dir", defaultConfDir, "main directory for bootstrap configuration")
-	cfgstruct.SetupFlag(zap.L(), rootCmd, &identityDir, "identity-dir", defaultIdentityDir, "main directory for bootstrap identity credentials")
-	defaults := cfgstruct.DefaultsFlag(rootCmd)
-	rootCmd.AddCommand(runCmd)
-	rootCmd.AddCommand(setupCmd)
-	process.Bind(runCmd, &runCfg, defaults, cfgstruct.ConfDir(confDir), cfgstruct.IdentityDir(identityDir))
-	process.Bind(setupCmd, &setupCfg, defaults, cfgstruct.ConfDir(confDir), cfgstruct.IdentityDir(identityDir), cfgstruct.SetupMode())
-}
-
-func cmdRun(cmd *cobra.Command, args []string) (err error) {
-	// inert constructors only ====
-
-	ctx, _ := process.Ctx(cmd)
-	log := zap.L()
-
-	identity, err := runCfg.Identity.Load()
-	if err != nil {
-		zap.S().Fatal(err)
-	}
-
-	if err := runCfg.Verify(log); err != nil {
-		log.Sugar().Error("Invalid configuration: ", err)
-		return err
-	}
-
-	db, err := bootstrapdb.New(bootstrapdb.Config{
-		Kademlia: runCfg.Kademlia.DBPath,
-	})
-	if err != nil {
-		return errs.New("Error starting master database on bootstrap: %+v", err)
-	}
-
-	defer func() {
-		err = errs.Combine(err, db.Close())
-	}()
-
-	revocationDB, err := revocation.NewDBFromCfg(runCfg.Server.Config)
-	if err != nil {
-		return errs.New("Error creating revocation database: %+v", err)
-	}
-	defer func() {
-		err = errs.Combine(err, revocationDB.Close())
-	}()
-
-	peer, err := bootstrap.New(log, identity, db, revocationDB, runCfg, version.Build)
-	if err != nil {
-		return err
-	}
-
-	// okay, start doing stuff ====
-
-	err = peer.Version.CheckVersion(ctx)
-	if err != nil {
-		return err
-	}
-
-	if err := process.InitMetricsWithCertPath(ctx, log, nil, runCfg.Identity.CertPath); err != nil {
-		zap.S().Warn("Failed to initialize telemetry batcher: ", err)
-	}
-
-	err = db.CreateTables()
-	if err != nil {
-		return errs.New("Error creating tables for master database on bootstrap: %+v", err)
-	}
-
-	runError := peer.Run(ctx)
-	closeError := peer.Close()
-
-	return errs.Combine(runError, closeError)
-}
-
-func cmdSetup(cmd *cobra.Command, args []string) (err error) {
-	setupDir, err := filepath.Abs(confDir)
-	if err != nil {
-		return err
-	}
-
-	valid, _ := fpath.IsValidSetupDir(setupDir)
-	if !valid {
-		return fmt.Errorf("bootstrap configuration already exists (%v)", setupDir)
-	}
-
-	err = os.MkdirAll(setupDir, 0700)
-	if err != nil {
-		return err
-	}
-
-	overrides := map[string]interface{}{}
-
-	serverAddress := cmd.Flag("server.address")
-	if !serverAddress.Changed {
-		overrides[serverAddress.Name] = defaultServerAddr
-	}
-
-	kademliaBootstrapAddr := cmd.Flag("kademlia.bootstrap-addr")
-	if !kademliaBootstrapAddr.Changed {
-		overrides[kademliaBootstrapAddr.Name] = "127.0.0.1" + defaultServerAddr
-	}
-
-	return process.SaveConfig(cmd, filepath.Join(setupDir, "config.yaml"),
-		process.SaveConfigWithOverrides(overrides))
-}
-
-func main() {
-	process.Exec(rootCmd)
-}
