+++ cmd/satellite/entrypoint
@@ -19,18 +19,10 @@ if [[ -n "${IDENTITY_ADDR:-}" ]]; then
 	export STORJ_SERVER_ADDRESS="${IDENTITY_ADDR}"
 fi
 
-if [[ -n "${BOOTSTRAP_ADDR:-}" ]]; then
-	export STORJ_KADEMLIA_BOOTSTRAP_ADDR="${BOOTSTRAP_ADDR}"
-fi
-
 if [[ ! -f "${CONF_PATH}/config.yaml" ]]; then
 	./satellite setup $SETUP_PARAMS
 fi
 
 RUN_PARAMS="${RUN_PARAMS:-} --config-dir ${CONF_PATH}"
-
-if [[ -n "${BOOTSTRAP_ADDR:-}" ]]; then
-	RUN_PARAMS="${RUN_PARAMS} --kademlia.bootstrap-addr ${BOOTSTRAP_ADDR}"
-fi
 
 exec ./satellite run $RUN_PARAMS "$@"
