+++ cmd/storagenode/storagenode.yaml
@@ -45,8 +45,6 @@ spec:
             value: "7777"
           - name: KAD_PORT
             value: "8080"
-          - name: KAD_HOST
-            value: "bootstrap.storj.io"
           - name: KAD_LISTEN_PORT
             value: "7776"
           - name: PS_DIR
