+++ internal/testplanet/reconfigure.go
@@ -8,16 +8,12 @@ import (
 
 	"go.uber.org/zap"
 
-	"storj.io/storj/bootstrap"
 	"storj.io/storj/satellite"
 	"storj.io/storj/storagenode"
 )
 
 // Reconfigure allows to change node configurations
 type Reconfigure struct {
-	NewBootstrapDB func(index int) (bootstrap.DB, error)
-	Bootstrap      func(index int, config *bootstrap.Config)
-
 	NewSatelliteDB func(log *zap.Logger, index int) (satellite.DB, error)
 	Satellite      func(log *zap.Logger, index int, config *satellite.Config)
 
@@ -29,9 +25,6 @@ type Reconfigure struct {
 // DisablePeerCAWhitelist returns a `Reconfigure` that sets `UsePeerCAWhitelist` for
 // all node types that use kademlia.
 var DisablePeerCAWhitelist = Reconfigure{
-	Bootstrap: func(index int, config *bootstrap.Config) {
-		config.Server.UsePeerCAWhitelist = false
-	},
 	Satellite: func(log *zap.Logger, index int, config *satellite.Config) {
 		config.Server.UsePeerCAWhitelist = false
 	},
