+++ pkg/pb/node.proto
@@ -26,7 +26,7 @@ enum NodeType {
     SATELLITE = 1;
     STORAGE = 2;
     UPLINK = 3;
-    BOOTSTRAP = 4;
+    BOOTSTRAP = 4; //TODO remove
 }
 
 // NodeAddress contains the information needed to communicate with a node on the network
