+++ versioncontrol/peer.go
@@ -27,7 +27,6 @@ type Config struct {
 
 // ServiceVersions provides a list of allowed Versions per Service
 type ServiceVersions struct {
-	Bootstrap   string `user:"true" help:"Allowed Bootstrap Versions" default:"v0.0.1"`
 	Satellite   string `user:"true" help:"Allowed Satellite Versions" default:"v0.0.1"`
 	Storagenode string `user:"true" help:"Allowed Storagenode Versions" default:"v0.0.1"`
 	Uplink      string `user:"true" help:"Allowed Uplink Versions" default:"v0.0.1"`
@@ -101,11 +100,6 @@ func New(log *zap.Logger, config *Config) (peer *Peer, err error) {
 	}
 
 	// Convert each Service's Version String to SemVer
-	peer.Versions.Bootstrap, err = version.NewSemVer(config.Versions.Bootstrap)
-	if err != nil {
-		return &Peer{}, err
-	}
-
 	peer.Versions.Satellite, err = version.NewSemVer(config.Versions.Satellite)
 	if err != nil {
 		return &Peer{}, err
