--- web/bootstrap/.gitignore
@@ -1,26 +1,0 @@
-.DS_Store
-node_modules
-dist
-package-lock.json
-coverage
-
-# local env files
-.env.local
-.env.*.local
-
-# Log files
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# Allow images
-!*.svg
-
-# Editor directories and files
-.idea
-.vscode
-*.suo
-*.ntvs*
-*.njsproj
-*.sln
-*.sw*
