--- web/bootstrap/babel.config.js
@@ -1,8 +1,0 @@
-// Copyright (C) 2019 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-module.exports = {
-  presets: [
-    '@vue/app'
-  ]
-}
