--- web/bootstrap/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, user-scalable=no">
-        <title>Storj Node Bootstrap</title>
-    </head>
-    <body>
-        <div id="app"></div>
-    </body>
-</html>
