--- web/bootstrap/package.json
@@ -1,65 +1,0 @@
-{
-  "name": "storj-bootstrap",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "serve": "rm -rf dist/ && npm run build && vue-cli-service serve -s dist",
-    "lint": "vue-cli-service lint",
-    "build": "webpack --config webpack.config.dev.js"
-  },
-  "dependencies": {
-    "@types/graphql": "^14.0.3",
-    "apollo-cache-inmemory": "^1.3.9",
-    "apollo-client": "^2.4.5",
-    "apollo-link": "^1.2.4",
-    "apollo-link-context": "^1.0.10",
-    "apollo-link-http": "^1.5.5",
-    "graphql": "^14.0.2",
-    "graphql-tag": "^2.10.0",
-    "vue": "^2.5.17",
-    "vue-apollo": "^3.0.0-beta.25",
-    "vue-class-component": "^6.0.0",
-    "vue-clipboards": "^1.2.4",
-    "vue-property-decorator": "^7.0.0",
-    "vue-router": "^3.0.1",
-    "vuex": "^3.0.1"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
-    "@types/sinon": "^5.0.5",
-    "@vue/cli-plugin-babel": "^3.0.5",
-    "@vue/cli-plugin-typescript": "^3.0.5",
-    "@vue/cli-plugin-unit-jest": "^3.0.5",
-    "@vue/cli-service": "^3.0.5",
-    "babel-core": "7.0.0-bridge.0",
-    "css-loader": "^1.0.0",
-    "eslint": "^5.9.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "node-sass": "^4.9.0",
-    "sass-loader": "^7.0.1",
-    "tslint": "^5.11.0",
-    "tslint-consistent-codestyle": "^1.14.1",
-    "tslint-loader": "^3.5.4",
-    "typescript": "^3.0.0",
-    "vue-html-webpack-plugin": "^3.2.2",
-    "vue-loader": "^15.4.2",
-    "vue-style-loader": "^4.1.2",
-    "vue-template-compiler": "^2.5.17",
-    "vue-tslint": "^0.3.2",
-    "vue-tslint-loader": "^3.5.6",
-    "webpack": "^4.21.0",
-    "webpack-cli": "^3.1.2",
-    "webpack-node-externals": "^1.7.2"
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not ie <= 8"
-  ]
-}
