--- web/bootstrap/src/main.ts
@@ -1,15 +1,0 @@
-// Copyright (C) 2019 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-import Vue from 'vue';
-import App from './App.vue';
-import router from './router';
-import store from './store';
-
-Vue.config.productionTip = false;
-
-new Vue({
-    router,
-    render: (h) => h(App),
-    store,
-}).$mount('#app');
