--- web/bootstrap/src/types/nodeStatus.ts
@@ -1,8 +1,0 @@
-// Copyright (C) 2019 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-export enum NodeStatus {
-    None = 1,
-    Active,
-    Error,
-}
