--- web/bootstrap/src/utils/constants.ts
@@ -1,12 +1,0 @@
-// Copyright (C) 2019 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-export const MUTATIONS = {
-    SET_NODE_STATUS: 'SET_NODE_STATUS',
-    SET_LOADING: 'SET_LOADING',
-};
-
-export const ACTIONS = {
-    CHECK_NODE_STATUS: 'CHECK_NODE_STATUS',
-    SET_LOADING: 'SET_LOADING',
-};
