--- web/bootstrap/src/views/Failure.vue
@@ -1,162 +1,0 @@
-// Copyright (C) 2019 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-<template>
-	<div class="failure-container">
-        <svg id="loading-svg" width="122px" height="120px" viewBox="0 0 122 120" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
-            <defs>
-                <linearGradient x1="62.8427859%" y1="49.4456705%" x2="-91.4339241%" y2="51.3777925%" id="linearGradient-1">
-                    <stop stop-color="#FF0000" offset="0%"></stop>
-                    <stop stop-color="#505460" offset="100%"></stop>
-                </linearGradient>
-                <filter x="-35.1%" y="-1516.0%" width="170.2%" height="3132.0%" filterUnits="objectBoundingBox" id="filter-2">
-                    <feGaussianBlur stdDeviation="8" in="SourceGraphic"></feGaussianBlur>
-                </filter>
-                <linearGradient x1="62.8427859%" y1="32.6019784%" x2="-91.4339241%" y2="93.2429871%" id="linearGradient-3">
-                    <stop stop-color="#FF0000" offset="0%"></stop>
-                    <stop stop-color="#505460" offset="100%"></stop>
-                </linearGradient>
-                <filter x="-213.9%" y="-1516.0%" width="527.8%" height="3132.0%" filterUnits="objectBoundingBox" id="filter-4">
-                    <feGaussianBlur stdDeviation="8" in="SourceGraphic"></feGaussianBlur>
-                </filter>
-                <linearGradient x1="62.8427859%" y1="-400.886306%" x2="-91.4339241%" y2="1170.68321%" id="linearGradient-5">
-                    <stop stop-color="#FF0000" offset="0%"></stop>
-                    <stop stop-color="#505460" offset="100%"></stop>
-                </linearGradient>
-                <filter x="-38.2%" y="-54.9%" width="176.4%" height="209.8%" filterUnits="objectBoundingBox" id="filter-6">
-                    <feGaussianBlur stdDeviation="8" in="SourceGraphic"></feGaussianBlur>
-                </filter>
-                <linearGradient x1="62.8427859%" y1="-424.051899%" x2="-91.4339241%" y2="1228.26156%" id="linearGradient-7">
-                    <stop stop-color="#FF0000" offset="0%"></stop>
-                    <stop stop-color="#505460" offset="100%"></stop>
-                </linearGradient>
-                <linearGradient x1="62.8427859%" y1="-11.1496077%" x2="-91.4339241%" y2="201.988068%" id="linearGradient-8">
-                    <stop stop-color="#FF0000" offset="0%"></stop>
-                    <stop stop-color="#505460" offset="100%"></stop>
-                </linearGradient>
-            </defs>
-            <g id="Icon/node/2-Copy-2" stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
-                <g id="Group-2" transform="translate(23.000000, 20.000000)">
-                    <g id="Group" opacity="0.870117188" transform="translate(3.000000, 13.000000)" stroke-width="4.2">
-                        <path d="M5.68434189e-14,25.0148333 L68.3666667,25.0148333" id="Stroke-3" stroke="url(#linearGradient-1)" opacity="0.605352493" filter="url(#filter-2)"></path>
-                        <path d="M58.553,35.6936111 L46.3496667,35.6936111" id="Stroke-5" stroke="url(#linearGradient-3)" opacity="0.605352493" stroke-linecap="round" filter="url(#filter-4)"></path>
-                        <path d="M68.3666667,25.1483333 L68.3666667,34.2016667 C68.3666667,41.565 62.3983333,47.5333333 55.035,47.5333333 L13.3316667,47.5333333 C5.96833333,47.5333333 5.68434189e-14,41.565 5.68434189e-14,34.2016667 L5.68434189e-14,25.015 C5.68434189e-14,24.4783333 0.128333333,23.9516667 0.376666667,23.4766667 L8.855,7.17833333 C11.1516667,2.76666667 15.7116667,-2.34479103e-13 20.685,-2.34479103e-13 L45.945,-2.34479103e-13 C50.5583333,-2.34479103e-13 54.8466667,2.385 57.2766667,6.30833333 L67.8666667,23.3916667 C68.1933333,23.92 68.3666667,24.5283333 68.3666667,25.1483333 Z" id="Stroke-1" stroke="url(#linearGradient-5)" opacity="0.605352493" filter="url(#filter-6)"></path>
-                    </g>
-                    <g id="Group" fill-rule="nonzero">
-                        <g id="Group-4" transform="translate(0.000000, 13.000000)">
-                            <g id="Group-3">
-                                <g id="Group-7">
-                                    <g id="Group-5" transform="translate(0.250333, 0.250833)" fill="url(#linearGradient-7)">
-                                        <path d="M6.03606851,25.9148333 L68.1003134,25.9148333 L58.4913685,10.4140957 C56.4451258,7.11035405 52.8329636,5.1 48.945,5.1 L23.685,5.1 C19.4945489,5.1 15.6526847,7.43113725 13.7180055,11.147462 L6.03606851,25.9148333 Z M69.2666667,30.1148333 L5.1,30.1148333 L5.1,37.2016667 C5.1,43.405202 10.1281313,48.4333333 16.3316667,48.4333333 L58.035,48.4333333 C64.2385354,48.4333333 69.2666667,43.405202 69.2666667,37.2016667 L69.2666667,30.1148333 Z M73.4666667,28.1483333 L73.4666667,37.2016667 C73.4666667,45.724798 66.5581313,52.6333333 58.035,52.6333333 L16.3316667,52.6333333 C7.80853536,52.6333333 0.9,45.724798 0.9,37.2016667 L0.9,28.015 C0.9,27.1392671 1.11042376,26.2788214 1.51366113,25.507538 L9.99229426,9.20862857 C12.6504846,4.1025112 17.9284291,0.9 23.685,0.9 L48.945,0.9 C54.2867182,0.9 59.2497936,3.66220673 62.0615403,8.20188581 L72.6528236,25.287292 C73.1845977,26.1473551 73.4666667,27.1380876 73.4666667,28.1483333 Z" id="Combined-Shape"></path>
-                                    </g>
-                                    <path d="M61.8033333,36.8444444 C62.9631313,36.8444444 63.9033333,37.7846465 63.9033333,38.9444444 C63.9033333,40.1042424 62.9631313,41.0444444 61.8033333,41.0444444 L49.6,41.0444444 C48.440202,41.0444444 47.5,40.1042424 47.5,38.9444444 C47.5,37.7846465 48.440202,36.8444444 49.6,36.8444444 L61.8033333,36.8444444 Z" id="Stroke-5" fill="url(#linearGradient-8)"></path>
-                                </g>
-                            </g>
-                        </g>
-                        <path d="M88.2734342,15.145147 C89.1803944,16.0521073 89.1803944,17.4586211 88.2734342,18.3655813 C87.8199541,18.8190614 87.2300672,19.0458015 86.6862538,19.0458015 C86.1424405,19.0458015 85.50648,18.8190614 85.0990735,18.3655813 L79.4768272,12.743335 L73.8545809,18.3197798 C73.4011008,18.7732599 72.8112138,19 72.2674005,19 C71.7235872,19 71.0876267,18.7732599 70.6802202,18.3197798 C69.7732599,17.4128196 69.7732599,16.0063058 70.6802202,15.0993456 L76.256665,9.52290075 L70.6802202,3.90065444 C69.7732599,2.99369424 69.7732599,1.58718036 70.6802202,0.680220153 C71.5871804,-0.226740051 72.9936942,-0.226740051 73.9006544,0.680220153 L79.4770993,6.30246645 L85.0993456,0.680220153 C86.0063058,-0.226740051 87.4128196,-0.226740051 88.3197798,0.680220153 C89.2267401,1.58718036 89.2267401,2.99369424 88.3197798,3.90065444 L82.6975335,9.52290075 L88.2734342,15.145147 Z" id="Path" fill="#FF0000"></path>
-                    </g>
-                </g>
-            </g>
-		</svg>
-        <div class="failure-container__animation-container">
-            <p class="failure-container__title">Your Node Not Found</p>
-            <p class="failure-container__info">Please check your Node ID or restart your Storj Node and try again later</p>
-            <div class="overlay__main-container__button" @click="onTryAgainClick">
-                <p class="overlay__main-container__button__text">Try Again</p>
-            </div>
-            <p class="overlay__main-container__support">Or Contact our <a>Support</a></p>
-        </div>
-	</div>
-</template>
-
-<script lang="ts">
-import { Component, Vue } from 'vue-property-decorator';
-
-@Component({
-    mounted() {
-        setTimeout(() => {
-            (document as any).querySelector('.failure-container').classList.add('active');
-        }, 500);
-    },
-    methods: {
-        onTryAgainClick: function () {
-            location.reload();
-        },
-    },
-})
-
-export default class Failure extends Vue {}
-</script>
-
-<style lang="scss">
-    .failure-container {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        flex-direction: column;
-
-        &__title {
-            width: 558px;
-            font-family: 'font_bold';
-            font-size: 48px;
-            font-style: normal;
-            font-stretch: normal;
-            line-height: 57px;
-            letter-spacing: normal;
-            text-align: center;
-            color: #fefeff;
-            margin-block-start: 0em;
-            margin-block-end: 0em;
-        }
-
-        &__info {
-            margin-top: 24px;
-            width: 446px;
-            font-family: 'font_regular';
-            font-size: 16px;
-            font-style: normal;
-            font-stretch: normal;
-            line-height: 23px;
-            letter-spacing: normal;
-            text-align: center;
-            color: #696c77;
-            margin-block-start: 1.5em;
-            margin-block-end: 0em;
-        }
-
-        &__animation-container {
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            flex-direction: column;
-            visibility: hidden;
-            position: relative;
-            bottom: -200px;
-            opacity: 0;
-            -webkit-transition: all 0.5s linear;
-            -moz-transition: all 0.5s linear;
-            -o-transition: all 0.5s linear;
-            transition: all 0.5s linear;
-        }
-
-        svg {
-            display: block;
-            position: relative;
-            top: 192px;
-            transition: all 0.5s linear;
-        }
-    }
-
-    .failure-container.active {
-        svg {
-            top: 0;
-        }
-
-        .failure-container__animation-container {
-            margin-bottom: 0;
-            bottom: 0;
-            visibility: visible;
-            opacity: 1;
-        }
-    }
-</style>
