--- web/bootstrap/src/views/Success.vue
@@ -1,154 +1,0 @@
-// Copyright (C) 2019 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-<template>
-    <div class="success-container" >
-        <svg width="122px" height="120px" viewBox="0 0 122 120" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
-            <defs>
-                <linearGradient x1="62.8427859%" y1="49.4456705%" x2="-91.4339241%" y2="51.3777925%" id="linearGradient-1">
-                    <stop stop-color="#0058FF" offset="0%"></stop>
-                    <stop stop-color="#505460" offset="100%"></stop>
-                </linearGradient>
-                <filter x="-35.1%" y="-1516.0%" width="170.2%" height="3132.0%" filterUnits="objectBoundingBox" id="filter-2">
-                    <feGaussianBlur stdDeviation="8" in="SourceGraphic"></feGaussianBlur>
-                </filter>
-                <linearGradient x1="62.8427859%" y1="32.6019784%" x2="-91.4339241%" y2="93.2429871%" id="linearGradient-3">
-                    <stop stop-color="#0058FF" offset="0%"></stop>
-                    <stop stop-color="#505460" offset="100%"></stop>
-                </linearGradient>
-                <filter x="-213.9%" y="-1516.0%" width="527.8%" height="3132.0%" filterUnits="objectBoundingBox" id="filter-4">
-                    <feGaussianBlur stdDeviation="8" in="SourceGraphic"></feGaussianBlur>
-                </filter>
-                <linearGradient x1="62.8427859%" y1="-400.886306%" x2="-91.4339241%" y2="1170.68321%" id="linearGradient-5">
-                    <stop stop-color="#0058FF" offset="0%"></stop>
-                    <stop stop-color="#505460" offset="100%"></stop>
-                </linearGradient>
-                <filter x="-38.2%" y="-54.9%" width="176.4%" height="209.8%" filterUnits="objectBoundingBox" id="filter-6">
-                    <feGaussianBlur stdDeviation="8" in="SourceGraphic"></feGaussianBlur>
-                </filter>
-                <linearGradient x1="62.8427859%" y1="46.4797858%" x2="-91.4339241%" y2="58.7495336%" id="linearGradient-7">
-                    <stop stop-color="#4381F7" offset="0%"></stop>
-                    <stop stop-color="#044CFF" offset="100%"></stop>
-                </linearGradient>
-                <linearGradient x1="62.8427859%" y1="-424.051899%" x2="-91.4339241%" y2="1228.26156%" id="linearGradient-8">
-                    <stop stop-color="#4381F7" offset="0%"></stop>
-                    <stop stop-color="#044CFF" offset="100%"></stop>
-                </linearGradient>
-                <linearGradient x1="62.8427859%" y1="-11.1496077%" x2="-91.4339241%" y2="201.988068%" id="linearGradient-9">
-                    <stop stop-color="#4381F7" offset="0%"></stop>
-                    <stop stop-color="#505460" offset="100%"></stop>
-                </linearGradient>
-            </defs>
-            <g id="Icon/node/2-Copy" stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
-                <g id="Group-2" transform="translate(23.000000, 21.000000)">
-                    <g id="Group" transform="translate(3.000000, 18.000000)" stroke-width="4.2">
-                        <path d="M5.68434189e-14,25.0148333 L68.3666667,25.0148333" id="Stroke-3" stroke="url(#linearGradient-1)" filter="url(#filter-2)"></path>
-                        <path d="M58.553,35.6936111 L46.3496667,35.6936111" id="Stroke-5" stroke="url(#linearGradient-3)" stroke-linecap="round" filter="url(#filter-4)"></path>
-                        <path d="M68.3666667,25.1483333 L68.3666667,34.2016667 C68.3666667,41.565 62.3983333,47.5333333 55.035,47.5333333 L13.3316667,47.5333333 C5.96833333,47.5333333 5.68434189e-14,41.565 5.68434189e-14,34.2016667 L5.68434189e-14,25.015 C5.68434189e-14,24.4783333 0.128333333,23.9516667 0.376666667,23.4766667 L8.855,7.17833333 C11.1516667,2.76666667 15.7116667,-2.34479103e-13 20.685,-2.34479103e-13 L45.945,-2.34479103e-13 C50.5583333,-2.34479103e-13 54.8466667,2.385 57.2766667,6.30833333 L67.8666667,23.3916667 C68.1933333,23.92 68.3666667,24.5283333 68.3666667,25.1483333 Z" id="Stroke-1" stroke="url(#linearGradient-5)" filter="url(#filter-6)"></path>
-                    </g>
-                    <g id="Group-4" transform="translate(0.250333, 12.250833)" fill-rule="nonzero">
-                        <g id="Group-3">
-                            <g id="Group-5" transform="translate(0.900000, 0.900000)">
-                                <polygon id="Stroke-3" fill="url(#linearGradient-7)" points="2.1 29.2148333 2.1 25.0148333 70.4666667 25.0148333 70.4666667 29.2148333"></polygon>
-                                <path d="M68.3666667,27.2483333 C68.3666667,27.0190576 68.3023215,26.7930533 68.181793,26.5981142 L57.5913685,9.51409574 C55.5451258,6.21035405 51.9329636,4.2 48.045,4.2 L22.785,4.2 C18.5945489,4.2 14.7526847,6.53113725 12.8180055,10.247462 L4.33767917,26.5496171 C4.24711487,26.7228441 4.2,26.9155019 4.2,27.115 L4.2,36.3016667 C4.2,42.505202 9.22813131,47.5333333 15.4316667,47.5333333 L57.135,47.5333333 C63.3385354,47.5333333 68.3666667,42.505202 68.3666667,36.3016667 L68.3666667,27.2483333 Z M72.5666667,27.2483333 L72.5666667,36.3016667 C72.5666667,44.824798 65.6581313,51.7333333 57.135,51.7333333 L15.4316667,51.7333333 C6.90853536,51.7333333 -2.84217094e-14,44.824798 -2.84217094e-14,36.3016667 L-2.84217094e-14,27.115 C-2.84217094e-14,26.2392671 0.210423755,25.3788214 0.613661128,24.607538 L9.09229426,8.30862857 C11.7504846,3.2025112 17.0284291,0 22.785,0 L48.045,0 C53.3867182,0 58.3497936,2.76220673 61.1615403,7.30188581 L71.7528236,24.387292 C72.2845977,25.2473551 72.5666667,26.2380876 72.5666667,27.2483333 Z" id="Stroke-1" fill="url(#linearGradient-8)"></path>
-                            </g>
-                        </g>
-                        <path d="M61.553,36.5936111 C62.712798,36.5936111 63.653,37.5338131 63.653,38.6936111 C63.653,39.8534091 62.712798,40.7936111 61.553,40.7936111 L49.3496667,40.7936111 C48.1898687,40.7936111 47.2496667,39.8534091 47.2496667,38.6936111 C47.2496667,37.5338131 48.1898687,36.5936111 49.3496667,36.5936111 L61.553,36.5936111 Z" id="Stroke-5" fill="url(#linearGradient-9)"></path>
-                    </g>
-                    <path d="M69.44417,8.21779453 L69.44417,8.21779453 C70.2538459,7.26705512 71.6659377,7.11616 72.6582279,7.87434261 L78.4,12.2614787 L88.9509915,1.59935829 C89.8468741,0.694039749 91.2857828,0.622683594 92.26684,1.43492382 L92.26684,1.43492382 C93.1329574,2.1520027 93.2537769,3.43543684 92.536698,4.30155419 C92.4934859,4.35374756 92.4476951,4.40375115 92.3994967,4.45137826 L78.68834,18 L69.7763018,11.2574595 C68.850355,10.5569201 68.6676262,9.23839261 69.3681655,8.31244584 C69.392584,8.28017041 69.4179294,8.24860682 69.44417,8.21779453 Z" id="Path" fill="#5FC000" fill-rule="nonzero"></path>
-                </g>
-            </g>
-        </svg>
-        <div class="success-container__animation-container">
-            <h1 class="success-container__title">Connected</h1>
-            <h3 class="success-container__info">Your Node found and it is connected to the network</h3>
-        </div>
-    </div>
-</template>
-
-<script lang="ts">
-    import { Component, Vue } from 'vue-property-decorator';
-
-    @Component({
-        mounted() {
-            setTimeout(() => {
-                (document as any).querySelector('.success-container').classList.add('active');
-            }, 500);
-        }
-    })
-
-    export default class Success extends Vue {}
-</script>
-
-<style scoped lang="scss">
-    .success-container {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        flex-direction: column;
-
-        &__title {
-            width: 278px;
-            font-family: 'font_bold';
-            font-size: 48px;
-            font-style: normal;
-            font-stretch: normal;
-            line-height: 57px;
-            letter-spacing: normal;
-            text-align: center;
-            color: #fefeff;
-            margin-block-start: 0em;
-            margin-block-end: 0em;
-        }
-
-        &__info {
-            margin-top: 24px;
-            width: 446px;
-            font-family: 'font_regular';
-            font-size: 16px;
-            font-style: normal;
-            font-stretch: normal;
-            line-height: 23px;
-            letter-spacing: normal;
-            text-align: center;
-            color: #696c77;
-            margin-block-start: 1.5em;
-            margin-block-end: 0em;
-        }
-
-        &__animation-container {
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            flex-direction: column;
-            visibility: hidden;
-            position: relative;
-            bottom: -50px;
-            opacity: 0;
-            -webkit-transition: all 0.5s linear;
-            -moz-transition: all 0.5s linear;
-            -o-transition: all 0.5s linear;
-            transition: all 0.5s linear;
-        }
-
-        svg {
-            display: block;
-            position: relative;
-            top: 50px;
-            transition: all 0.5s linear;
-        }
-    }
-
-    .success-container.active {
-        svg {
-            top: 0;
-        }
-
-        .success-container__animation-container {
-            margin-bottom: 0;
-            bottom: 0;
-            visibility: visible;
-            opacity: 1;
-        }
-    }
-</style>
