+++ web/marketing/pages/base.html
@@ -9,7 +9,7 @@ See LICENSE for copying information. -->
         </head>
         <body>
             {{template "body" .}}
-            {{template "bootstrap" .}}
+            {{template "bootstrap" .}} /*TODO necessary?*/
         </body>
     </html>
 {{end}}
